clean up internals
