+++ source/config/Config.ts
@@ -9,20 +9,7 @@ import { JsonScanner } from "./JsonScanner.js";
 import { OptionGroup } from "./OptionGroup.enum.js";
 import { Target } from "./Target.js";
 import { defaultOptions } from "./defaultOptions.js";
-import type { CommandLineOptions, ConfigFileOptions, OptionValue } from "./types.js";
-
-export interface ResolvedConfig
-  extends Omit<CommandLineOptions, "config" | keyof ConfigFileOptions>,
-    Required<ConfigFileOptions> {
-  /**
-   * The path to a TSTyche configuration file.
-   */
-  configFilePath: string;
-  /**
-   * Only run test files with matching path.
-   */
-  pathMatch: Array<string>;
-}
+import type { CommandLineOptions, ConfigFileOptions, OptionValue, ResolvedConfig } from "./types.js";
 
 export class Config {
   static #onDiagnostics(this: void, diagnostic: Diagnostic) {
