+++ source/config/index.ts
@@ -1,8 +1,8 @@
 export { ConfigDiagnosticText } from "./ConfigDiagnosticText.js";
-export { Config, type ResolvedConfig } from "./Config.js";
-export { Directive, type Directives } from "./Directive.js";
+export { Config } from "./Config.js";
+export { Directive } from "./Directive.js";
 export { OptionBrand } from "./OptionBrand.enum.js";
 export { type ItemDefinition, type OptionDefinition, Options } from "./Options.js";
 export { OptionGroup } from "./OptionGroup.enum.js";
 export { defaultOptions } from "./defaultOptions.js";
-export type { CommandLineOptions, ConfigFileOptions } from "./types.js";
+export type { CommandLineOptions, ConfigFileOptions, InlineConfig, ResolvedConfig } from "./types.js";
